pyproject.toml
[tool.poetry]
name = "nondeterministic-agent"
version = "0.1.0"
description = ""
authors = ["Yuriy Babyak <yuriy.babyak@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
litellm = "^1.63.11"
colorlog = "^6.9.0"
prompt-toolkit = "^3.0.50"
aider-chat = "^0.78.0"
httpx = "^0.28.0"
pydantic = "^2.10.0"
pyyaml = "^6.0.2"
python-dotenv = "1.0.1"
pydantic-settings = "^2.8.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
